+++ tests/VersionSmokeTest.php
@@ -6,7 +6,7 @@ require __DIR__ . '/bootstrap.php';
 
 use Mnb\PHPExcel\MnbExcel;
 
-smoke_assert_equals('1.2.0', MnbExcel::VERSION, 'VERSION constant should match this release');
-smoke_assert_equals('1.2.0', MnbExcel::version(), 'version() should match this release');
+smoke_assert_equals('1.3.0', MnbExcel::VERSION, 'VERSION constant should match this release');
+smoke_assert_equals('1.3.0', MnbExcel::version(), 'version() should match this release');
 
 echo "VersionSmokeTest passed\n";